@@ -16,23 +16,22 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.0.0",
+    "@angular/animations": "^9.1.13",
     "@angular/cdk": "^8.0.0",
-    "@angular/common": "^8.0.0",
-    "@angular/compiler": "^8.0.0",
-    "@angular/compiler-cli": "^8.0.0",
-    "@angular/core": "^8.0.0",
-    "@angular/forms": "^8.0.0",
-    "@angular/platform-browser": "^8.0.0",
-    "@angular/platform-browser-dynamic": "^8.0.0",
-    "@angular/router": "^8.0.0",
+    "@angular/common": "^9.1.13",
+    "@angular/compiler": "^9.1.13",
+    "@angular/compiler-cli": "^9.1.13",
+    "@angular/core": "^9.1.13",
+    "@angular/forms": "^9.1.13",
+    "@angular/platform-browser": "^9.1.13",
+    "@angular/platform-browser-dynamic": "^9.1.13",
+    "@angular/router": "^9.1.13",
     "@babel/polyfill": "^7.12.1",
     "@fortawesome/fontawesome-svg-core": "^1.2.18",
     "@fortawesome/free-brands-svg-icons": "^5.8.2",
     "@fortawesome/free-solid-svg-icons": "^5.8.2",
     "@ngx-share/core": "^7.1.2",
     "@types/file-saver": "^2.0.1",
-    "@types/webappsec-credential-management": "^0.6.1",
     "angular-datatables": "^8.1.0",
     "angular2-ladda": "^2.0.1",
     "angular2-recaptcha": "1.1.0",
@@ -47,20 +46,20 @@
     "ngx-bootstrap": "^4.2.0",
     "rxjs": "^6.5.2",
     "rxjs-compat": "^6.5.2",
-    "tinymce": "^4.9.4",
+    "tinymce": "^4.9.11",
     "ts-helpers": "^1.1.2",
-    "tslib": "^1.9.0",
-    "typescript": "3.4.5",
-    "zone.js": "~0.9.1"
+    "tslib": "^1.10.0",
+    "typescript": "3.8.3",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.800.0",
-    "@angular/cli": "^8.0.1",
+    "@angular-devkit/build-angular": "~0.901.15",
+    "@angular/cli": "^9.1.15",
     "@types/datatables.net": "^1.10.17",
     "@types/jasmine": "3.3.13",
     "@types/jquery": "^3.3.29",
-    "@types/node": "~12.0.4",
-    "codelyzer": "^5.0.1",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "^3.6.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.1.0",
