@@ -56,7 +56,7 @@ import {WindowRef} from "../../services/window-ref";
               </div>
           </div>
           <div class="form-group box-footer">
-              <button [disabled]="loading" class="btn btn-primary" [ladda]="loading">{{'btn.login' | translate}}
+              <button type="submit" [disabled]="loading" class="btn btn-primary" [ladda]="loading">{{'btn.login' | translate}}
               </button>
           </div>
       </form>
@@ -76,7 +76,6 @@ export class LoginComponent implements OnInit {
   loading = false;
   error = false;
   lastUserName = 'lastUserName';
-  storeCreds = false;
 
   constructor(
       private winRef: WindowRef,
@@ -88,10 +87,6 @@ export class LoginComponent implements OnInit {
   ngOnInit() {
     this.model.username = this.getLastUserName();
     this.model.rememberUserName = !isNullOrUndefined(this.model.username);
-    if (this.winRef.nativeWindow.PasswordCredential || this.winRef.nativeWindow.FederatedCredential) {
-      this.storeCreds = true;
-      this.winRef.nativeWindow.navigator.credentials.get();
-    }
   }
 
   login() {
@@ -101,20 +96,6 @@ export class LoginComponent implements OnInit {
         .subscribe(
             isLoggedIn => {
               if (isLoggedIn) {
-                //store using browsers creds api
-                if (this.storeCreds) {
-
-                  let credential = new PasswordCredential({
-                    id: this.model.username, // Username/ID
-                    password: this.model.password // Password
-                  });
-
-                  this.winRef.nativeWindow.navigator.credentials.store(credential).then(() => {
-                    console.info("Credential stored in the user agent's credential manager.");
-                  }, (err) => {
-                    console.error("Error while storing the credential: ", err);
-                  });
-                }
                 //On success, save username
                 this.setLastUserName(this.model.rememberUserName ? this.model.username : undefined);
                 //redirect if needed
